+++ app/gui2/src/stores/graph.ts
@@ -341,7 +341,7 @@ export const useGraphStore = defineStore('graph', () => {
 
   return {
     _parsed,
-    _parsedEnso: _parsedEnso,
+    _parsedEnso,
     transact,
     nodes,
     exprNodes,
