+++ app/gui2/src/util/computedValueRegistry.ts
@@ -1,12 +1,12 @@
-import type { ExecutionContext } from '@/stores/project.ts'
-import { reactive } from 'vue'
+import type { ExecutionContext } from '@/stores/project'
 import type {
   ExpressionId,
   ExpressionUpdate,
   ExpressionUpdatePayload,
   MethodCall,
   ProfilingInfo,
-} from '../../shared/languageServerTypes.ts'
+} from 'shared/languageServerTypes'
+import { reactive } from 'vue'
 
 export interface ExpressionInfo {
   typename: string | undefined
