+++ app/gui2/src/util/measurement.ts
