+++ app/gui2/src/util/visualizationDataRegistry.ts
@@ -1,13 +1,13 @@
 import type { ExecutionContext } from '@/stores/project.ts'
 import { OutboundPayload, VisualizationUpdate } from 'shared/binaryProtocol.ts'
 import type { DataServer } from 'shared/dataServer.ts'
-import { reactive } from 'vue'
 import type {
   ExpressionUpdatePayload,
   MethodCall,
   ProfilingInfo,
   Uuid,
-} from '../../shared/languageServerTypes.ts'
+} from 'shared/languageServerTypes.ts'
+import { reactive } from 'vue'
 
 export interface ExpressionInfo {
   typename: string | undefined
