+++ app/gui2/src/views/ProjectView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import GraphEditor from '../components/GraphEditor.vue'
+import GraphEditor from '@/components/GraphEditor.vue'
 </script>
 
 <template>
